RF move compcor methods to a new module to reduce length
